RAC-367 refactor : 패키지 변경
