+++ src/main/java/com/postgraduate/domain/mentoring/application/usecase/MentoringApplyingUseCase.java
@@ -13,8 +13,8 @@ import com.postgraduate.domain.payment.domain.entity.Payment;
 import com.postgraduate.domain.payment.domain.service.PaymentGetService;
 import com.postgraduate.domain.senior.domain.entity.Senior;
 import com.postgraduate.domain.user.domain.entity.User;
-import com.postgraduate.global.bizppurio.usecase.BizppurioJuniorMessage;
-import com.postgraduate.global.bizppurio.usecase.BizppurioSeniorMessage;
+import com.postgraduate.global.bizppurio.application.usecase.BizppurioJuniorMessage;
+import com.postgraduate.global.bizppurio.application.usecase.BizppurioSeniorMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
