+++ src/main/java/com/postgraduate/global/bizppurio/application/dto/req/ContentRequest.java
@@ -1,0 +1,5 @@
+package com.postgraduate.global.bizppurio.application.dto.req;
+
+import com.postgraduate.global.bizppurio.application.dto.req.content.Message;
+
+public record ContentRequest(Message at) {}
