+++ src/main/java/com/postgraduate/global/bizppurio/application/dto/req/content/CertificationDeniedMessage.java
@@ -1,6 +1,6 @@
-package com.postgraduate.global.bizppurio.dto.req.content;
+package com.postgraduate.global.bizppurio.application.dto.req.content;
 
-import com.postgraduate.global.bizppurio.dto.req.content.button.WebLinkButton;
+import com.postgraduate.global.bizppurio.application.dto.req.content.button.WebLinkButton;
 
 public record CertificationDeniedMessage(
         String message,
