+++ src/main/java/com/postgraduate/global/bizppurio/application/dto/req/content/Message.java
@@ -1,0 +1,4 @@
+package com.postgraduate.global.bizppurio.application.dto.req.content;
+
+public interface Message {
+}
