+++ src/main/java/com/postgraduate/global/bizppurio/application/dto/res/BizppurioTokenResponse.java
@@ -1,3 +1,3 @@
-package com.postgraduate.global.bizppurio.dto.res;
+package com.postgraduate.global.bizppurio.application.dto.res;
 
 public record BizppurioTokenResponse(String accesstoken, String type, String expired) {}
