+++ src/main/java/com/postgraduate/global/bizppurio/application/usecase/BizppurioSeniorMessage.java
@@ -1,8 +1,8 @@
-package com.postgraduate.global.bizppurio.usecase;
+package com.postgraduate.global.bizppurio.application.usecase;
 
 import com.postgraduate.domain.senior.domain.entity.Senior;
 import com.postgraduate.domain.user.domain.entity.User;
-import com.postgraduate.global.bizppurio.mapper.BizppurioMapper;
+import com.postgraduate.global.bizppurio.application.mapper.BizppurioMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
