--- src/main/java/com/postgraduate/global/bizppurio/dto/req/ContentRequest.java
@@ -1,5 +1,0 @@
-package com.postgraduate.global.bizppurio.dto.req;
-
-import com.postgraduate.global.bizppurio.dto.req.content.Message;
-
-public record ContentRequest(Message at) {}
