--- src/main/java/com/postgraduate/global/bizppurio/dto/req/content/Message.java
@@ -1,4 +1,0 @@
-package com.postgraduate.global.bizppurio.dto.req.content;
-
-public interface Message {
-}
