+++ src/test/java/com/postgraduate/domain/auth/application/usecase/oauth/SignUpUseTypeTest.java
@@ -19,7 +19,7 @@ import com.postgraduate.domain.user.exception.PhoneNumberException;
 import com.postgraduate.domain.wish.domain.entity.Wish;
 import com.postgraduate.domain.wish.domain.entity.constant.Status;
 import com.postgraduate.domain.wish.domain.service.WishSaveService;
-import com.postgraduate.global.bizppurio.usecase.BizppurioSeniorMessage;
+import com.postgraduate.global.bizppurio.application.usecase.BizppurioSeniorMessage;
 import com.postgraduate.global.slack.SlackSignUpMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
