+++ src/test/java/com/postgraduate/domain/mentoring/application/usecase/MentoringApplyingUseTypeTest.java
@@ -17,8 +17,8 @@ import com.postgraduate.domain.senior.domain.entity.Info;
 import com.postgraduate.domain.senior.domain.entity.Profile;
 import com.postgraduate.domain.senior.domain.entity.Senior;
 import com.postgraduate.domain.user.domain.entity.User;
-import com.postgraduate.global.bizppurio.usecase.BizppurioJuniorMessage;
-import com.postgraduate.global.bizppurio.usecase.BizppurioSeniorMessage;
+import com.postgraduate.global.bizppurio.application.usecase.BizppurioJuniorMessage;
+import com.postgraduate.global.bizppurio.application.usecase.BizppurioSeniorMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
